analysis_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
elasticsearch==8.11.0
fluent-logger==0.10.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
kafka-python==2.0.2
aiokafka==0.8.1
aiohttp==3.9.1
pyjwt==2.10.1

# Тестовые зависимости
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-mock==3.12.0 